requirements.txt
# Python packages for multimodal embedding pipeline
# Install on Databricks cluster

# Core embedding models
sentence-transformers==2.3.1
transformers==4.36.2
torch==2.1.2

# CLIP model dependencies
ftfy==6.1.3
regex==2023.12.25

# Image processing
Pillow==10.2.0
opencv-python==4.9.0.80

# Azure OpenAI (optional, for future use)
openai==1.10.0

# Utilities
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
